Extract artifact normalization helper
