+++ php-transformer/src/ArtifactCompiler/ArtifactCompiler.php
@@ -8,16 +8,13 @@ use Automattic\BlocksEngine\PhpTransformer\Contract\TransformerResult;
 final class ArtifactCompiler
 {
     private const INPUT_SCHEMA = 'block-artifact-compiler/website-artifact/v1';
-    private const DEFAULT_MAX_FILES = 500;
-    private const DEFAULT_MAX_FILE_BYTES = 1048576;
-    private const DEFAULT_MAX_TOTAL_BYTES = 10485760;
 
     /**
      * @param array<string, mixed> $artifact
      */
     public function compile(array $artifact): TransformerResult
     {
-        $normalized = $this->normalizeArtifact($artifact);
+        $normalized = ( new ArtifactNormalizer() )->normalize($artifact);
         $entry = $this->entryFile($normalized['files'], $normalized['entrypoints']);
         $documents = $this->compileSourceDocuments($normalized);
         $diagnostics = array_merge($normalized['diagnostics'], $documents['diagnostics']);
@@ -51,9 +48,9 @@ final class ArtifactCompiler
                 'files_by_source' => $this->countBy($normalized['files'], 'source'),
                 'files_by_intent' => $this->countBy($normalized['files'], 'intent'),
                 'limits'          => array(
-                    'max_files'       => self::DEFAULT_MAX_FILES,
-                    'max_file_bytes'  => self::DEFAULT_MAX_FILE_BYTES,
-                    'max_total_bytes' => self::DEFAULT_MAX_TOTAL_BYTES,
+                    'max_files'       => ArtifactNormalizer::DEFAULT_MAX_FILES,
+                    'max_file_bytes'  => ArtifactNormalizer::DEFAULT_MAX_FILE_BYTES,
+                    'max_total_bytes' => ArtifactNormalizer::DEFAULT_MAX_TOTAL_BYTES,
                 ),
                 'source_hash'     => hash('sha256', $normalized['hash_payload']),
                 'html'            => array(
@@ -92,134 +89,6 @@ final class ArtifactCompiler
                     'source_hash'   => hash('sha256', $normalized['hash_payload']),
                 ),
             )
-        );
-    }
-
-    /**
-     * @param array<string, mixed> $artifact
-     * @return array{files: array<int, array<string, mixed>>, diagnostics: array<int, array<string, mixed>>, rejected_count: int, bytes: int, entrypoints: array<int, string>, hash_payload: string}
-     */
-    private function normalizeArtifact(array $artifact): array
-    {
-        $diagnostics = array();
-        $files = array();
-        $entrypoints = array();
-        $rejected = 0;
-        $bytes = 0;
-        $seenPaths = array();
-
-        foreach ( array('entrypoint', 'entry', 'main') as $key ) {
-            if ( is_string($artifact[$key] ?? null) ) {
-                $entrypoints[] = $artifact[$key];
-            }
-        }
-        if ( is_array($artifact['entrypoints'] ?? null) ) {
-            foreach ( $artifact['entrypoints'] as $entrypoint ) {
-                if ( is_string($entrypoint) ) {
-                    $entrypoints[] = $entrypoint;
-                }
-            }
-        }
-
-        $rawFiles = $this->rawFiles($artifact);
-        $safeEntrypoints = array();
-        foreach ( array_unique($entrypoints) as $entrypoint ) {
-            $path = $this->safeRelativePath($entrypoint);
-            if ( '' === $path ) {
-                $diagnostics[] = $this->diagnostic('unsafe_entrypoint_path', 'warning', 'An artifact entrypoint was ignored because its path is empty, absolute, or escapes the artifact root.', array('path' => $entrypoint));
-                continue;
-            }
-            $safeEntrypoints[] = $path;
-        }
-
-        foreach ( $rawFiles as $index => $file ) {
-            if ( count($files) >= self::DEFAULT_MAX_FILES ) {
-                ++$rejected;
-                $diagnostics[] = $this->diagnostic('file_limit_exceeded', 'warning', 'Additional artifact files were ignored because the file limit was reached.', array('max_files' => self::DEFAULT_MAX_FILES));
-                break;
-            }
-
-            $path = $this->safeRelativePath((string) ($file['path'] ?? ''));
-            if ( '' === $path ) {
-                ++$rejected;
-                $diagnostics[] = $this->diagnostic('unsafe_artifact_path', 'warning', 'An artifact file was ignored because its path is empty, absolute, or escapes the artifact root.', array('index' => $index));
-                continue;
-            }
-
-            $payload = $this->payload($file, $path);
-            $diagnostics = array_merge($diagnostics, $payload['diagnostics']);
-            if ( ! $payload['accepted'] ) {
-                ++$rejected;
-                continue;
-            }
-
-            if ( $payload['bytes'] > self::DEFAULT_MAX_FILE_BYTES ) {
-                ++$rejected;
-                $diagnostics[] = $this->diagnostic('artifact_file_too_large', 'warning', 'An artifact file was ignored because it exceeds the per-file byte limit.', array('path' => $path, 'bytes' => $payload['bytes'], 'max_file_bytes' => self::DEFAULT_MAX_FILE_BYTES));
-                continue;
-            }
-
-            if ( $bytes + $payload['bytes'] > self::DEFAULT_MAX_TOTAL_BYTES ) {
-                ++$rejected;
-                $diagnostics[] = $this->diagnostic('artifact_total_too_large', 'warning', 'An artifact file was ignored because the bundle byte limit was reached.', array('path' => $path, 'bytes' => $payload['bytes'], 'max_total_bytes' => self::DEFAULT_MAX_TOTAL_BYTES));
-                continue;
-            }
-
-            $path = $this->dedupePath($path, $seenPaths);
-            $seenPaths[$path] = true;
-            $mimeType = $this->mimeType((string) ($file['mime_type'] ?? $file['mime'] ?? $file['media_type'] ?? (str_contains((string) ($file['type'] ?? ''), '/') ? $file['type'] : '')), $path);
-            $kind = $this->kind((string) ($file['kind'] ?? $file['type'] ?? ''), $path, $payload['content'], $mimeType);
-            $role = $this->role((string) ($file['role'] ?? ''), $kind, $mimeType, $path);
-            $intent = $this->intent((string) ($file['intent'] ?? ''), $kind, $role);
-            $binary = $payload['binary'] || $this->isBinaryMimeType($mimeType);
-            $contentBase64 = $payload['content_base64'];
-            if ( $binary && '' === $contentBase64 ) {
-                $contentBase64 = base64_encode($payload['content']);
-            }
-            $entrypoint = in_array($path, $safeEntrypoints, true) || ! empty($file['entrypoint']) || 'entry' === $role;
-            if ( $entrypoint && ! in_array($path, $safeEntrypoints, true) ) {
-                $safeEntrypoints[] = $path;
-            }
-
-            $normalized = array(
-                'path'       => $path,
-                'content'    => $payload['content'],
-                'kind'       => $kind,
-                'bytes'      => $payload['bytes'],
-                'source'     => (string) ($file['source'] ?? 'artifact'),
-                'mime_type'  => $mimeType,
-                'role'       => $role,
-                'encoding'   => $payload['encoding'],
-                'binary'     => $binary,
-                'entrypoint' => $entrypoint,
-                'provenance' => array(
-                    'source_path' => $path,
-                    'source'      => (string) ($file['source'] ?? 'artifact'),
-                    'hash'        => hash('sha256', '' !== $contentBase64 ? $contentBase64 : $payload['content']),
-                ),
-            );
-            if ( '' !== $contentBase64 ) {
-                $normalized['content_base64'] = $contentBase64;
-            }
-            if ( '' !== $intent ) {
-                $normalized['intent'] = $intent;
-            }
-
-            if ( 'mdx' === $kind ) {
-                $diagnostics[] = $this->diagnostic('mdx_source_document_detected', 'warning', 'MDX source document support is partial; the source was preserved and inspectable document/component metadata was extracted.', array('path' => $path));
-            }
-
-            $bytes += $normalized['bytes'];
-            $files[] = $normalized;
-        }
-
-        return array(
-            'files'          => $files,
-            'diagnostics'    => $this->dedupeDiagnostics($diagnostics),
-            'rejected_count' => $rejected,
-            'bytes'          => $bytes,
-            'entrypoints'    => array_values(array_unique($safeEntrypoints)),
-            'hash_payload'   => $this->fileHashPayload($files),
         );
     }
 
@@ -302,111 +171,6 @@ final class ArtifactCompiler
     }
 
     /**
-     * @param array<string, mixed> $artifact
-     * @return array<int, array<string, mixed>>
-     */
-    private function rawFiles(array $artifact): array
-    {
-        $files = array();
-        foreach ( array('files', 'artifacts', 'outputs') as $key ) {
-            if ( ! is_array($artifact[$key] ?? null) ) {
-                continue;
-            }
-            foreach ( $artifact[$key] as $path => $file ) {
-                if ( is_array($file) ) {
-                    $pathSource = $file['path'] ?? $file['name'] ?? $path;
-                    $file['path'] = is_scalar($pathSource) ? (string) $pathSource : '';
-                    $file['source'] = is_scalar($file['source'] ?? null) ? (string) $file['source'] : $key;
-                    $files[] = $file;
-                    continue;
-                }
-                if ( is_string($file) ) {
-                    $files[] = array(
-                        'path'    => is_string($path) ? $path : 'artifact-' . $path . '.html',
-                        'content' => $file,
-                        'kind'    => '',
-                        'source'  => $key,
-                    );
-                }
-            }
-        }
-        foreach ( array('html', 'generated_html', 'content', 'body') as $key ) {
-            if ( is_string($artifact[$key] ?? null) && '' !== trim($artifact[$key]) ) {
-                $files[] = array(
-                    'path'    => 'index.html',
-                    'content' => $artifact[$key],
-                    'kind'    => 'html',
-                    'source'  => $key,
-                );
-            }
-        }
-        foreach ( array(
-            'css'        => 'style.css',
-            'styles'     => 'style.css',
-            'javascript' => 'site.js',
-            'js'         => 'site.js',
-            'script'     => 'site.js',
-        ) as $key => $path ) {
-            if ( is_string($artifact[$key] ?? null) && '' !== trim($artifact[$key]) ) {
-                $files[] = array(
-                    'path'    => $path,
-                    'content' => $artifact[$key],
-                    'kind'    => str_contains($path, '.css') ? 'css' : 'js',
-                    'source'  => $key,
-                );
-            }
-        }
-
-        return $files;
-    }
-
-    /**
-     * @param array<string, mixed> $file
-     * @return array{accepted: bool, content: string, content_base64: string, encoding: string, binary: bool, bytes: int, diagnostics: array<int, array<string, mixed>>}
-     */
-    private function payload(array $file, string $path): array
-    {
-        if ( is_string($file['content_base64'] ?? null) ) {
-            $base64 = preg_replace('/\s+/', '', $file['content_base64']) ?? '';
-            $decoded = base64_decode($base64, true);
-            if ( false === $decoded ) {
-                return array('accepted' => false, 'content' => '', 'content_base64' => '', 'encoding' => 'base64', 'binary' => false, 'bytes' => 0, 'diagnostics' => array($this->diagnostic('invalid_base64_content', 'warning', 'An artifact file was ignored because content_base64 is not valid base64.', array('path' => $path))));
-            }
-
-            $binary = $this->looksBinary($decoded);
-            $diagnostics = array();
-            if ( ! $binary && is_string($file['content'] ?? null) && '' !== $file['content'] && $file['content'] !== $decoded ) {
-                $diagnostics[] = $this->diagnostic('content_base64_preferred', 'info', 'Both content and content_base64 were provided; decoded content_base64 was used as the canonical payload.', array('path' => $path));
-            }
-
-            return array('accepted' => true, 'content' => $binary ? '' : $decoded, 'content_base64' => $base64, 'encoding' => 'base64', 'binary' => $binary, 'bytes' => strlen($decoded), 'diagnostics' => $diagnostics);
-        }
-
-        $content = $this->normalizeContent($file['content'] ?? $file['body'] ?? $file['text'] ?? '');
-        return array('accepted' => true, 'content' => $content, 'content_base64' => '', 'encoding' => 'text', 'binary' => false, 'bytes' => strlen($content), 'diagnostics' => array());
-    }
-
-    private function safeRelativePath(string $path): string
-    {
-        $path = str_replace('\\', '/', trim($path));
-        if ( '' === $path || str_starts_with($path, '/') || preg_match('#^[A-Za-z]:/#', $path) ) {
-            return '';
-        }
-        $parts = array();
-        foreach ( explode('/', $path) as $part ) {
-            if ( '' === $part || '.' === $part ) {
-                continue;
-            }
-            if ( '..' === $part ) {
-                return '';
-            }
-            $parts[] = $part;
-        }
-
-        return implode('/', $parts);
-    }
-
-    /**
      * @param array<int, array<string, mixed>> $files
      * @param array<int, string> $entrypoints
      * @return array<string, mixed>|null
@@ -434,157 +198,6 @@ final class ArtifactCompiler
         }
 
         return null;
-    }
-
-    private function kind(string $kind, string $path, string $content, string $mimeType): string
-    {
-        $kind = $this->sanitizeKey($kind);
-        if ( in_array($kind, array('html', 'css', 'js', 'jsx', 'tsx', 'json', 'markdown', 'mdx', 'asset', 'blocks'), true) ) {
-            return $kind;
-        }
-        if ( str_contains($mimeType, '/') ) {
-            if ( str_contains($mimeType, 'html') ) {
-                return 'html';
-            }
-            if ( 'text/css' === $mimeType ) {
-                return 'css';
-            }
-            if ( in_array($mimeType, array('application/javascript', 'text/javascript', 'application/ecmascript', 'text/ecmascript'), true) ) {
-                return 'js';
-            }
-            if ( 'application/json' === $mimeType ) {
-                return 'json';
-            }
-        }
-
-        $extension = strtolower(pathinfo($path, PATHINFO_EXTENSION));
-        return match ($extension) {
-            'html', 'htm' => 'html',
-            'css' => 'css',
-            'js', 'mjs' => 'js',
-            'jsx' => 'jsx',
-            'tsx' => 'tsx',
-            'json' => 'json',
-            'md', 'markdown' => 'markdown',
-            'mdx' => 'mdx',
-            default => str_contains($content, '<!-- wp:') ? 'blocks' : 'asset',
-        };
-    }
-
-    private function mimeType(string $mimeType, string $path): string
-    {
-        $mimeType = strtolower(trim($mimeType));
-        if ( preg_match('#^[a-z0-9.+-]+/[a-z0-9.+-]+$#', $mimeType) ) {
-            return $mimeType;
-        }
-        return match (strtolower(pathinfo($path, PATHINFO_EXTENSION))) {
-            'html', 'htm' => 'text/html',
-            'css' => 'text/css',
-            'js', 'mjs' => 'application/javascript',
-            'jsx' => 'text/jsx',
-            'tsx' => 'text/tsx',
-            'json' => 'application/json',
-            'md', 'markdown' => 'text/markdown',
-            'mdx' => 'text/mdx',
-            'txt' => 'text/plain',
-            'svg' => 'image/svg+xml',
-            'png' => 'image/png',
-            'jpg', 'jpeg' => 'image/jpeg',
-            'gif' => 'image/gif',
-            'webp' => 'image/webp',
-            'avif' => 'image/avif',
-            'woff' => 'font/woff',
-            'woff2' => 'font/woff2',
-            'ttf' => 'font/ttf',
-            'otf' => 'font/otf',
-            default => 'application/octet-stream',
-        };
-    }
-
-    private function isBinaryMimeType(string $mimeType): bool
-    {
-        return ! str_starts_with($mimeType, 'text/') && ! in_array($mimeType, array('application/json', 'application/javascript', 'image/svg+xml'), true);
-    }
-
-    private function role(string $role, string $kind, string $mimeType, string $path): string
-    {
-        $role = $this->sanitizeKey($role);
-        if ( '' !== $role ) {
-            return $role;
-        }
-        if ( 'html' === $kind ) {
-            return preg_match('#(^|/)index\.html?$#i', $path) ? 'entry' : 'document';
-        }
-        if ( 'css' === $kind ) {
-            return 'stylesheet';
-        }
-        if ( 'js' === $kind ) {
-            return 'script';
-        }
-        if ( str_starts_with($mimeType, 'image/') ) {
-            return 'image';
-        }
-        if ( str_starts_with($mimeType, 'font/') ) {
-            return 'font';
-        }
-        if ( in_array($kind, array('json', 'markdown'), true) ) {
-            return 'data';
-        }
-
-        return 'asset';
-    }
-
-    private function intent(string $intent, string $kind, string $role): string
-    {
-        $intent = $this->sanitizeKey($intent);
-        if ( '' !== $intent ) {
-            return $intent;
-        }
-        if ( 'css' === $kind || 'stylesheet' === $role ) {
-            return 'style';
-        }
-        if ( 'js' === $kind || 'script' === $role ) {
-            return 'behavior';
-        }
-
-        return '';
-    }
-
-    private function looksBinary(string $content): bool
-    {
-        return str_contains($content, "\0");
-    }
-
-    private function normalizeContent(mixed $content): string
-    {
-        if ( is_string($content) ) {
-            return str_replace("\r\n", "\n", str_replace("\r", "\n", $content));
-        }
-        if ( is_scalar($content) ) {
-            return (string) $content;
-        }
-
-        return '';
-    }
-
-    /**
-     * @param array<string, bool> $seen
-     */
-    private function dedupePath(string $path, array $seen): string
-    {
-        if ( ! isset($seen[$path]) ) {
-            return $path;
-        }
-
-        $extension = pathinfo($path, PATHINFO_EXTENSION);
-        $base = '' === $extension ? $path : substr($path, 0, -1 - strlen($extension));
-        $suffix = '' === $extension ? '' : '.' . $extension;
-        $index = 2;
-        while ( isset($seen[$base . '-' . $index . $suffix]) ) {
-            ++$index;
-        }
-
-        return $base . '-' . $index . $suffix;
     }
 
     /**
